collapse page templates into PageTemplate; rename McparReportPage to ReportPage (#3739)
* text field

* form.tsx

* radio/checkbox

* date field

* dynamic field

* dynamic update

* number field

* label tweak

* dropdown field

* choice list field

* tweak

* single checkbox and cleanup

* fix failing tests

* add some hydration tests

* standardize field styles

* await form post before moving on

* fixed bug #2

* fix bug #4

* updated DateField tests

* better DateField mocks

* more tests

* whoops

* all tests passing. now to add more

* ChoiceList tests

* NumberField tests

* TextField tests

* clean up admin banner form

* clean up add edit program form

* combine mcpar form

* create and use flattened array

* clean up and pass around data

* add validation to form json

* add formschema to existing non-mcpar forms

* fix failing tests; add more

* custom hook for finding routes

* change variable name to clarify functionality

* modify repprtJson prop passing

* add ReportPageFooter tests

* move /mcpar/dashboard to /mcpar as root

* delete all the old separated forms

* clean up ADS

* last changes

* change BasicPage to PageTemplate

* combine ReportPage into PageTemplate

* migrate tests & delete ReportPage

* rename McparReportPage -> ReportPage

* cleanup and moving stuff

* add dropdown default option back in; update tests (#3742)
