+++ services/ui-src/src/components/app/Error.tsx
@@ -1,6 +1,6 @@
 // components
 import { Flex, Heading, Image, Link, Text } from "@chakra-ui/react";
-import { BasicPage } from "components";
+import { PageTemplate } from "components";
 // utils
 import { createEmailLink } from "utils/other/email";
 import { makeMediaQueryClasses } from "utils";
@@ -14,7 +14,7 @@ export const Error = () => {
   const { preLinkText, helpDeskEmail, postLinkText } = emailText;
 
   return (
-    <BasicPage data-testid="error-view" sxOverride={sx.layout}>
+    <PageTemplate data-testid="error-view" sxOverride={sx.layout}>
       <Flex sx={sx.heading}>
         <Image
           src={warningIcon}
@@ -36,7 +36,7 @@ export const Error = () => {
         </Link>
         {postLinkText}
       </Text>
-    </BasicPage>
+    </PageTemplate>
   );
 };
