+++ services/ui-src/src/routes/index.tsx
@@ -5,8 +5,7 @@ export { Help } from "./Help/Help";
 export { Home } from "./Home/Home";
 export { NotFound } from "./NotFound/NotFound";
 export { Profile } from "./Profile/Profile";
-// mcpar
+// report
 export { GetStarted } from "./GetStarted/GetStarted";
 export { Dashboard } from "./Dashboard/Dashboard";
-export { McparReportPage } from "./McparReportPage/McparReportPage";
 export { ReviewSubmit } from "./ReviewSubmit/ReviewSubmit";
